Apply code refactoring
