+++ src/components/Login/LoginCss.css
@@ -1,0 +1,108 @@
+body {
+	margin: 0;
+	padding: 0;
+	background-image: linear-gradient(rgba(0,0,0,0.7), rgba(0,0,0,0.7)),  url('../../assets/images/background.jpg');
+	background-size: cover;
+	height: 100vh;
+	overflow: hidden;
+}
+.center {
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	transform: translate(-50%, -50%);
+	width: 400px;
+	background: rgb(116 130 131);
+	border-radius: 10px;
+}
+.center h1{
+	text-align: center;
+	padding: 0 0 20px 0;
+	border-bottom: 1px solid #c0c0c0;
+	color: #000000;
+}
+.center form {
+	padding: 0 40px;
+	box-sizing: border-box;
+}
+form .txt_field{
+	position: relative;
+	border-bottom: 2px solid #000000;
+	margin: 30px 0;
+}
+.txt_field input {
+	width: 100%;
+	padding: 0 5px;
+	height: 40px;
+	font-size: 16px;
+	border: none;
+	background: none;
+	outline: none;
+}
+.txt_field label{
+	position: absolute;
+	top: 50%;
+	left: 5px;
+	color: #000000;
+	transform: translateY(-50%);
+	font-size: 16px;
+	pointer-events: none;
+	transition: .5s;
+}
+.txt_field span::before{
+	content: '';
+	position: absolute;
+	top: 40px;
+	left: 0;
+	width: 0%;
+	height: 2px;
+	background: #FFFFFF;
+	transition: .9s;
+}
+.txt_field input:focus ~ label,
+.txt_field input:valid ~ label{
+	top: -5px;
+	color: #FFFFFF;
+}
+.txt_field input:focus ~ span::before,
+.txt_field input:valid ~ span::before{
+  width: 100%;
+
+}
+.pass{
+	margin: -5px 0 20px 5px;
+	color: #a6a6a6;
+	cursor: pointer;
+}
+.pass a{
+  color: #000000;
+}
+.pass:hover{
+  text-decoration: underline;
+}
+input[type="submit"]{
+	width: 100%;
+	height: 50px;
+	border: 1px solid;
+	background: rgb(116 130 131);
+	border-radius: 25px;
+	font-size: 18px;
+	color: #000000;
+	font-weight: 700;
+	cursor: pointer;
+	outline: none;
+	margin-bottom: 50px;
+}
+input[type="submit"]:hover{
+	border-color: #FFFFFF;
+	color: #FFFFFF;
+	transition: 0.9s;
+}
+.close {
+	position: absolute;
+	right: 10px;
+	top: 0.05px;
+	font-size: 30px;
+	text-decoration: none;
+	color: inherit;
+}
